+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/base/BasePlugin.kt
@@ -27,7 +27,7 @@ class BasePlugin : Plugin {
     private var onMessage: OnMessage? = null
 
     @Transient
-    private var defaultPostRegisteredMessage: DefaultPostRegisteredMessage? = null
+    private var postMessagesRegistrant: PostMessagesRegistrant? = null
 
     @Transient
     val postsUsedTable = PostsUsedTable()
@@ -46,7 +46,7 @@ class BasePlugin : Plugin {
         onMediaGroup = OnMediaGroup(baseConfig.sourceChatId)
         onMessage = OnMessage(baseConfig.sourceChatId)
 
-        defaultPostRegisteredMessage = DefaultPostRegisteredMessage(
+        postMessagesRegistrant = PostMessagesRegistrant(
             executor,
             baseConfig.sourceChatId
         )